refactor: singleton theme detector and MarkdownContent wrapper
