+++ src/webview/src/components/Messages/MarkdownContent.vue
@@ -1,0 +1,18 @@
+<template>
+  <MarkdownRender
+    :content="content"
+    :enable-mermaid="true"
+    :is-dark="isDark"
+  />
+</template>
+
+<script setup lang="ts">
+import { MarkdownRender } from 'markstream-vue';
+import { useThemeDetector } from '../../utils/themeDetector';
+
+defineProps<{
+  content: string;
+}>();
+
+const { isDark } = useThemeDetector();
+</script>
