+++ src/webview/src/components/Messages/MermaidDiagram.vue
